cleanup: Move meteor implementation to ms-legacy directory
- Move MeteorShower component and rendering pipeline to ms-legacy/
- Remove meteor references from main codebase (app/page.tsx, debug config)
- Clean up WASM module exports and TypeScript interfaces
- Archive Phase 6 unified rendering implementation for future reference
- Main codebase now runs without meteor dependencies

This cleanup maintains the meteor shower work in ms-legacy/ while
simplifying the main application structure.
