+++ ms-legacy/lib/rendering/interfaces.ts
