+++ ms-legacy/lib/rendering/js-fallback/js-particle-system.ts
