+++ ms-legacy/lib/rendering/pipeline-factory.ts
