+++ ms-legacy/lib/rendering/typed-array-manager.ts
